# Conflicts:
#	phpstan-baseline.neon
#	src/Core/Framework/App/Lifecycle/Persister/CustomFieldPersister.php
#	src/Core/Framework/DependencyInjection/plugin.xml
#	src/Core/Framework/Plugin/PluginLifecycleService.php
#	tests/integration/Core/Framework/Plugin/KernelPluginIntegrationTest.php
#	tests/integration/Core/Framework/Plugin/PluginLifecycleServiceMigrationTest.php
#	tests/integration/Core/Framework/Plugin/PluginLifecycleServiceTest.php
#	tests/unit/Core/Framework/Plugin/PluginLifecycleServiceTest.php

+++ src/Elasticsearch/Admin/AdminSearchRegistry.php
@@ -6,7 +6,9 @@ use Doctrine\DBAL\ArrayParameterType;
 use Doctrine\DBAL\Connection;
 use Doctrine\DBAL\Exception;
 use OpenSearch\Client;
-use OpenSearch\Common\Exceptions\OpenSearchException;
+use OpenSearch\Exception\OpenSearchExceptionInterface;
+use Psr\Clock\ClockInterface;
+use Psr\Http\Client\ClientExceptionInterface;
 use Psr\Log\LoggerInterface;
 use Shopware\Core\Defaults;
 use Shopware\Core\Framework\Api\Context\SalesChannelApiSource;
@@ -14,7 +16,6 @@ use Shopware\Core\Framework\DataAbstractionLayer\Event\EntityWrittenContainerEve
 use Shopware\Core\Framework\Event\ProgressAdvancedEvent;
 use Shopware\Core\Framework\Event\ProgressFinishedEvent;
 use Shopware\Core\Framework\Event\ProgressStartedEvent;
-use Shopware\Core\Framework\Feature;
 use Shopware\Core\Framework\Log\Package;
 use Shopware\Core\Framework\Uuid\Uuid;
 use Shopware\Elasticsearch\Admin\Indexer\AbstractAdminIndexer;
@@ -54,7 +55,8 @@ class AdminSearchRegistry implements EventSubscriberInterface
         private readonly LoggerInterface $logger,
         array $config,
         private readonly array $mapping,
-        private readonly string $environment
+        private readonly string $environment,
+        private readonly ClockInterface $clock,
     ) {
         if (isset($config['settings']['index'])) {
             if (\array_key_exists('number_of_shards', $config['settings']['index']) && $config['settings']['index']['number_of_shards'] === null) {
@@ -92,7 +94,10 @@ class AdminSearchRegistry implements EventSubscriberInterface
         }
 
         $indexers = $this->getIndexersArray();
-        /** @var list<string> $entities */
+        if ($indexers === []) {
+            return;
+        }
+
         $entities = array_keys($indexers);
 
         if ($indexingBehavior->getOnlyEntities()) {
@@ -134,10 +139,15 @@ class AdminSearchRegistry implements EventSubscriberInterface
             return;
         }
 
+        $indexers = $this->getIndexersArray();
+        if ($indexers === []) {
+            return;
+        }
+
         if ($this->adminEsHelper->getRefreshIndices()) {
             try {
                 $this->refreshIndices();
-            } catch (OpenSearchException $e) {
+            } catch (ClientExceptionInterface|OpenSearchExceptionInterface $e) {
                 $this->logger->error('Could not refresh indices. Run "bin/console es:admin:mapping:update" & "bin/console es:admin:index" to update indices and reindex. Error: ' . $e->getMessage());
 
                 return;
@@ -146,12 +156,11 @@ class AdminSearchRegistry implements EventSubscriberInterface
 
         /** @var array<string, string> $indices */
         $indices = $this->connection->fetchAllKeyValue('SELECT `alias`, `index` FROM admin_elasticsearch_index_task');
-
-        if (empty($indices)) {
+        if ($indices === []) {
             return;
         }
 
-        foreach ($this->indexer as $indexer) {
+        foreach ($indexers as $indexer) {
             $ids = $indexer->getUpdatedIds($event);
             $deletedIds = $event->getDeletedPrimaryKeys($indexer->getEntity());
             $ids = array_values(array_diff($ids, $deletedIds));
@@ -250,7 +259,7 @@ class AdminSearchRegistry implements EventSubscriberInterface
             $documents[] = ['index' => ['_id' => $id]];
 
             $documents[] = \array_replace(
-                ['entityName' => $indexer->getEntity(), 'parameters' => [], 'textBoosted' => '', 'text' => ''],
+                ['entityName' => $indexer->getEntity(), 'parameters' => [], 'textBoosted' => '', 'text' => '', 'completion' => []],
                 $document
             );
         }
@@ -266,7 +275,7 @@ class AdminSearchRegistry implements EventSubscriberInterface
 
         $result = $this->client->bulk($arguments);
 
-        if (\is_array($result) && !empty($result['errors'])) {
+        if (\is_array($result) && ((bool) ($result['errors'] ?? false)) !== false) {
             $errors = $this->parseErrors($result);
 
             throw ElasticsearchException::indexingError($errors);
@@ -287,7 +296,7 @@ class AdminSearchRegistry implements EventSubscriberInterface
         foreach ($entities as $entityName) {
             $indexer = $this->getIndexer($entityName);
             $alias = $this->adminEsHelper->getIndex($indexer->getName());
-            $index = $alias . '_' . time();
+            $index = $alias . '_' . $this->clock->now()->getTimestamp();
 
             if ($this->client->indices()->exists(['index' => $index])) {
                 continue;
@@ -307,6 +316,10 @@ class AdminSearchRegistry implements EventSubscriberInterface
             ];
         }
 
+        if ($indices === []) {
+            return $indices;
+        }
+
         $this->connection->executeStatement(
             'DELETE FROM admin_elasticsearch_index_task WHERE `entity` IN (:entities)',
             ['entities' => $entities],
@@ -322,8 +335,8 @@ class AdminSearchRegistry implements EventSubscriberInterface
 
     private function refreshIndices(): void
     {
-        $entities = [];
         $indexTasks = [];
+        $entities = [];
         foreach ($this->indexer as $indexer) {
             $alias = $this->adminEsHelper->getIndex($indexer->getName());
 
@@ -331,7 +344,7 @@ class AdminSearchRegistry implements EventSubscriberInterface
                 continue;
             }
 
-            $index = $alias . '_' . time();
+            $index = $alias . '_' . $this->clock->now()->getTimestamp();
             $this->create($indexer, $index, $alias);
 
             $entities[] = $indexer->getEntity();
@@ -344,6 +357,10 @@ class AdminSearchRegistry implements EventSubscriberInterface
                 '`alias`' => $alias,
                 '`doc_count`' => $iterator->fetchCount(),
             ];
+        }
+
+        if ($entities === []) {
+            return;
         }
 
         $this->connection->executeStatement(
@@ -452,17 +469,13 @@ class AdminSearchRegistry implements EventSubscriberInterface
         $properties = [
             'properties' => [
                 'id' => AbstractElasticsearchDefinition::KEYWORD_FIELD,
-                'textBoosted' => AbstractAdminIndexer::SEARCH_FIELD,
-                'text' => AbstractAdminIndexer::SEARCH_FIELD,
+                'textBoosted' => AbstractAdminIndexer::TEXT_FIELD,
+                'text' => AbstractAdminIndexer::TEXT_FIELD,
+                'completion' => AbstractAdminIndexer::COMPLETION_FIELD,
                 'entityName' => AbstractElasticsearchDefinition::KEYWORD_FIELD,
                 'parameters' => AbstractElasticsearchDefinition::KEYWORD_FIELD,
             ],
         ];
-
-        if (Feature::isActive('ENABLE_OPENSEARCH_FOR_ADMIN_API')) {
-            $properties['properties']['textBoosted']['fields']['ngram']['search_analyzer'] = 'sw_whitespace_analyzer';
-            $properties['properties']['text']['fields']['ngram']['search_analyzer'] = 'sw_whitespace_analyzer';
-        }
 
         $mapping = $indexer->mapping($properties);
